Graphs are shown when calculation is done and when solar site is clicked, and deletion of solar sites is possible

+++ app/src/main/java/solarsitingucsc/smartsolarsiting/Controller/DisplayCalculationsActivity.java
@@ -21,9 +21,8 @@ import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.EditText;
-import android.widget.ListView;
 import android.widget.ProgressBar;
-import android.widget.TextView;
+import android.widget.Spinner;
 import android.widget.Toast;
 
 import com.android.volley.Request;
@@ -38,10 +37,10 @@ import com.github.mikephil.charting.components.XAxis;
 import com.github.mikephil.charting.data.Entry;
 import com.github.mikephil.charting.data.LineData;
 import com.github.mikephil.charting.data.LineDataSet;
+import com.github.mikephil.charting.formatter.IAxisValueFormatter;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 
-import com.github.mikephil.charting.formatter.IAxisValueFormatter;
 import com.google.api.client.extensions.android.json.AndroidJsonFactory;
 import com.google.api.client.http.javanet.NetHttpTransport;
 import com.google.api.services.vision.v1.Vision;
@@ -71,6 +70,7 @@ import org.json.JSONObject;
 import java.io.ByteArrayOutputStream;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
+import java.lang.reflect.Field;
 import java.text.DateFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -83,7 +83,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import java.util.Random;
 
 import solarsitingucsc.smartsolarsiting.Model.SolarSiting;
 import solarsitingucsc.smartsolarsiting.R;
@@ -96,14 +95,15 @@ public class DisplayCalculationsActivity extends AppCompatActivity {
     private double latitude, longitude;
     private final String DATASET_API_KEY = "iF9CgCZD45uP45g5ybzqYdvLINrToH60600nH9it";
     private final String GOOGLE_VISION_API_KEY = "AIzaSyDx2wu1igClYSoMYTfhvH5Mp0u5x9AxwrE";
+    private static final String[] months = {"January", "February", "March", "April", "May",
+            "June", "July", "August", "September", "October",
+            "November", "December"};
     private ProgressBar progressBar;
     private double[] hourlyArray = new double[8760];
-    private TextView[] textViews;
     private FirebaseAuth mAuth;
     private FloatingActionButton saveBtn;
     private String imageName;
     private String screenshotName;
-    private Bitmap imageToSave;
     private Bitmap thumbnail;
 
     @Override
@@ -118,57 +118,44 @@ public class DisplayCalculationsActivity extends AppCompatActivity {
                         finish();
                     }
                 });
-//        FloatingActionButton capture = findViewById(R.id.button_capture);
-//        capture.setOnClickListener(new View.OnClickListener() {
-//            @Override
-//            public void onClick(View view) {
-//                finish();
-//            }
-//        });
         saveBtn = findViewById(R.id.button_save);
         progressBar = findViewById(R.id.progressBar);
-        latitude = getIntent().getDoubleExtra("latitude", 0.0);
-        longitude = getIntent().getDoubleExtra("longitude", 0.0);
-        imageName = getIntent().getStringExtra("imageName");
-        screenshotName = getIntent().getStringExtra("screenshotName");
-        FileInputStream imageFis = null;
-        FileInputStream screenshotFis = null;
-        try {
-            imageFis = openFileInput(imageName);
-            screenshotFis = openFileInput(screenshotName);
-        } catch (FileNotFoundException e) {
-            Log.d(TAG, "File not found: " + e.getMessage());
+        HashMap<String, HashMap<String, Double>> powerList =
+                (HashMap<String, HashMap<String, Double>>) getIntent().getSerializableExtra("powerList");
+        if (powerList != null)
+            setupDropdown(powerList);
+        else {
+            latitude = getIntent().getDoubleExtra("latitude", 0.0);
+            longitude = getIntent().getDoubleExtra("longitude", 0.0);
+            imageName = getIntent().getStringExtra("imageName");
+            screenshotName = getIntent().getStringExtra("screenshotName");
+            FileInputStream imageFis = null;
+            FileInputStream screenshotFis = null;
+            try {
+                imageFis = openFileInput(imageName);
+                screenshotFis = openFileInput(screenshotName);
+            } catch (FileNotFoundException e) {
+                Log.d(TAG, "File not found: " + e.getMessage());
+            }
+            Bitmap originalImage = BitmapFactory.decodeStream(imageFis);
+            Matrix matrix = new Matrix();
+            matrix.postRotate(90);
+
+            Bitmap screenshot = BitmapFactory.decodeStream(screenshotFis);
+            thumbnail = makeThumbnail(originalImage, matrix);
+
+            new MakeGoogleRequest().execute(screenshot);
+            deleteFile(screenshotName);
         }
-        Bitmap originalImage = BitmapFactory.decodeStream(imageFis);
-        Matrix matrix = new Matrix();
-        matrix.postRotate(90);
-
-        imageToSave = Bitmap.createBitmap(originalImage, 0, 0, originalImage.getWidth(),
-                originalImage.getHeight(), matrix, true);
-        Bitmap screenshot = BitmapFactory.decodeStream(screenshotFis);
-        thumbnail = makeThumbnail(originalImage, matrix);
-
-        textViews = new TextView[13];
-        int[] months = {R.id.janKwTxt, R.id.febKwTxt, R.id.marKwTxt, R.id.aprKwTxt, R.id.mayKwTxt,
-                R.id.junKwTxt, R.id.julKwTxt, R.id.augKwTxt, R.id.sepKwTxt, R.id.octKwTxt,
-                R.id.novKwTxt, R.id.decKwTxt, R.id.annualKwTxt};
-        for (int i = 0; i < textViews.length; i++) {
-            textViews[i] = findViewById(months[i]);
-        }
-
-        new MakeGoogleRequest().execute(screenshot);
-        deleteFile(screenshotName);
-
-
-        new MakeGoogleRequest().execute(screenshot);
-
     }
 
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        deleteFile(imageName);
-        deleteFile(screenshotName);
+        if (imageName != null)
+            deleteFile(imageName);
+        if (screenshotName != null)
+            deleteFile(screenshotName);
     }
 
     private void makeDatasetRequest(double latitude, double longitude) {
@@ -426,7 +413,7 @@ public class DisplayCalculationsActivity extends AppCompatActivity {
         return "";
     }
 
-    private void storeResults(final String name, final int[] monthlyPower) {
+    private void storeResults(final String name, final HashMap<String, HashMap<String, Double>> monthlyPower) {
         final String id = getCurrentUserId();
         DatabaseReference rootRef = FirebaseDatabase.getInstance().getReference("pictures");
         rootRef.addListenerForSingleValueEvent(new ValueEventListener() {
@@ -439,10 +426,7 @@ public class DisplayCalculationsActivity extends AppCompatActivity {
                     storeImage(thumbnail, id + name);
                     Date today = Calendar.getInstance().getTime();
                     DateFormat df = DateFormat.getDateTimeInstance();
-                    SolarSiting solarSiting = new SolarSiting(
-                            id, name,
-                            Arrays.toString(monthlyPower).split("[\\[\\]]")[1].split(", "),
-                            id, df.format(today));
+                    SolarSiting solarSiting = new SolarSiting(id, name, monthlyPower, df.format(today));
                     solarSiting.store();
                     Toast.makeText(DisplayCalculationsActivity.this,
                             "Saved!", Toast.LENGTH_SHORT).show();
@@ -476,6 +460,104 @@ public class DisplayCalculationsActivity extends AppCompatActivity {
             public void onSuccess(UploadTask.TaskSnapshot taskSnapshot) {
                 // taskSnapshot.getMetadata() contains file metadata such as size, content-type, and download URL.
                 Uri downloadUrl = taskSnapshot.getDownloadUrl();
+            }
+        });
+    }
+
+    private void displayChart(String month, HashMap<String, HashMap<String, Double>> powerMap) {
+        LineChart chart = findViewById(R.id.chart);
+        chart.setVisibility(View.VISIBLE);
+        List<Entry> entries = new ArrayList<>();
+        XAxis xAxis = chart.getXAxis();
+        xAxis.setPosition(XAxis.XAxisPosition.BOTTOM);
+        xAxis.setDrawGridLines(false);
+        if (month.equals("All")) {
+            Double[] monthValues = new Double[12];
+            HashMap<String, Double> allValues = powerMap.get(month);
+            for (int i = 0; i < monthValues.length; i++) {
+                if (allValues.get(months[i]) != null && allValues.get(months[i]) instanceof Double)
+                    monthValues[i] = allValues.get(months[i]);
+                else
+                    monthValues[i] = 0d;
+            }
+            for (int i = 0; i < monthValues.length; i++) {
+                entries.add(new Entry(i, monthValues[i].floatValue()));
+            }
+            xAxis.setValueFormatter(new IAxisValueFormatter() {
+                @Override
+                public String getFormattedValue(float value, AxisBase axis) {
+                    return months[(int) value];
+                }
+            });
+        } else {
+            Double[] hours = new Double[24];
+            HashMap<String, Double> monthValues = powerMap.get(month);
+            for (int i = 0; i < hours.length; i++) {
+                if (monthValues != null && monthValues.get("" + i) != null)
+                    hours[i] = monthValues.get("" + i);
+                else
+                    hours[i] = 0d;
+            }
+            for (int i = 0; i < hours.length; i++) {
+                entries.add(new Entry(i, hours[i].floatValue()));
+            }
+            String[] stringHours = new String[hours.length];
+            for (int i = 0; i < stringHours.length; i++) {
+                stringHours[i] = i + ":00";
+            }
+            final String[] finalHours = stringHours;
+            xAxis.setValueFormatter(new IAxisValueFormatter() {
+                @Override
+                public String getFormattedValue(float value, AxisBase axis) {
+                    return finalHours[(int) value];
+                }
+            });
+        }
+        LineDataSet dataSet = new LineDataSet(entries, "Power in KW");
+        LineData lineData = new LineData(dataSet);
+        chart.setData(lineData);
+        chart.invalidate();
+    }
+
+    private void setupDropdown(final HashMap<String, HashMap<String, Double>> powerByTimeAndMonth) {
+        final Spinner dropdown = findViewById(R.id.spinner);
+
+        try {
+            Field popup = Spinner.class.getDeclaredField("mPopup");
+            popup.setAccessible(true);
+
+            // Get private mPopup member variable and try cast to ListPopupWindow
+            android.widget.ListPopupWindow popupWindow = (android.widget.ListPopupWindow) popup.get(dropdown);
+
+            // Set popupWindow height to 500px
+            popupWindow.setHeight(1000);
+        } catch (NoClassDefFoundError | ClassCastException | NoSuchFieldException | IllegalAccessException e) {
+            // silently fail...
+        }
+
+        // Initializing a new String Array
+        Resources res = getResources();
+
+        // Create a List from String Array elements
+        List<String> months = new ArrayList<>(Arrays.asList(res.getStringArray(R.array.months)));
+
+        // Create a ArrayAdapter from List
+        ArrayAdapter<String> arrayAdapter = new ArrayAdapter<>
+                (getApplicationContext(), android.R.layout.simple_spinner_dropdown_item, months);
+
+        // Populate spinner with items from ArrayAdapter
+        dropdown.setAdapter(arrayAdapter);
+
+        dropdown.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
+            @Override
+            public void onItemSelected(AdapterView<?> adapterView, View view, int i, long l) {
+                String month = dropdown.getSelectedItem().toString();
+                displayChart(month, powerByTimeAndMonth);
+            }
+
+            @Override
+            public void onNothingSelected(AdapterView<?> adapterView) {
+                //do nothing
             }
         });
     }
@@ -539,8 +621,10 @@ public class DisplayCalculationsActivity extends AppCompatActivity {
         protected void onPostExecute(String response) {
             Map<String, Integer> result = translateResponseToMap(response);
             Iterator it = result.entrySet().iterator();
-            int annualPower = 0;
-            final int[] monthlyPower = new int[13];
+            Double annualPower = 0d;
+            final Double[] monthlyPower = new Double[12];
+            Arrays.fill(monthlyPower, 0d);
+            final HashMap<String, HashMap<String, Double>> powerByTimeAndMonth = new HashMap<>();
             while (it.hasNext()) {
                 Map.Entry pair = (Map.Entry)it.next();
                 String key = (String) pair.getKey();
@@ -557,15 +641,33 @@ public class DisplayCalculationsActivity extends AppCompatActivity {
                         double power = getPowerForMonthAndHour(monthInt, hourInt)/6;
                         annualPower += power;
                         monthlyPower[monthInt] += power;
+                        HashMap<String, Double> newMap = powerByTimeAndMonth.get(months[monthInt]);
+                        if (newMap == null) {
+                            HashMap<String, Double> n = new HashMap<>();
+                            n.put("" + hourInt, power);
+                            powerByTimeAndMonth.put(months[monthInt], n);
+                        } else {
+                            if (newMap.get("" + hourInt) == null) {
+                                powerByTimeAndMonth.get(months[monthInt]).put("" + hourInt, power);
+                            } else {
+                                powerByTimeAndMonth.get(months[monthInt]).put("" + hourInt, newMap.get("" + hourInt) + power);
+                            }
+                        }
                     } catch (NumberFormatException e) {
                         break;
                     }
                 }
                 it.remove(); // avoids a ConcurrentModificationException
             }
-            monthlyPower[monthlyPower.length - 1] = annualPower;
+            HashMap<String, Double> t = new HashMap<>();
+            t.put("0", annualPower);
+            powerByTimeAndMonth.put("Annual", t);
+            HashMap<String, Double> temp = new HashMap<>();
+            for (int i = 0; i < monthlyPower.length; i++) {
+                temp.put(months[i], monthlyPower[i]);
+            }
+            powerByTimeAndMonth.put("All", temp);
             progressBar.setVisibility(View.GONE);
-
 
             saveBtn.setVisibility(View.VISIBLE);
             saveBtn.setOnClickListener(new View.OnClickListener() {
@@ -595,7 +697,7 @@ public class DisplayCalculationsActivity extends AppCompatActivity {
                                 dialog.cancel();
                             }
                             name[0] = input.getText().toString();
-                            storeResults(name[0], monthlyPower);
+                            storeResults(name[0], powerByTimeAndMonth);
                         }
                     });
                     builder.setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
@@ -607,42 +709,8 @@ public class DisplayCalculationsActivity extends AppCompatActivity {
                     builder.show();
                 }
             });
-
-            // Get reference of widgets from XML layout
-            ListView lv = (ListView) findViewById(R.id.months_list_view);
-
-            // Initializing a new String Array
-            Resources res = getResources();
-
-
-            // Create a List from String Array elements
-            List<String> months = new ArrayList<String>(Arrays.asList(res.getStringArray(R.array.months)));
-
-            // Create a ArrayAdapter from List
-            ArrayAdapter<String> arrayAdapter = new ArrayAdapter<String>
-                    (getApplicationContext(), android.R.layout.simple_list_item_1, months);
-
-            // Populate ListView with items from ArrayAdapter
-            lv.setAdapter(arrayAdapter);
-
-            // Set an item click listener for ListView
-            lv.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-                @Override
-                public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                    // Get the selected item text from ListView
-                    String selectedItem = (String) parent.getItemAtPosition(position);
-                    Intent intent = new Intent(getApplicationContext(), DisplayPlotActivity.class);
-                    startActivity(intent);
-
-                    intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-
-
-                }
-            });
-
-
+            setupDropdown(powerByTimeAndMonth);
         }
-
     }
 }
